refactor Tools/serial: move non-serial param generation into separate script
