+++ Tools/serial/serial_params.c.jinja
@@ -71,5 +71,3 @@ PARAM_DEFINE_INT32({{ command.port_param_name }}, {{ command.default_port }});
 {% endfor -%}
 {% endif %}
 
-{{ additional_definitions }}
-
